requirements.txt
beautifulsoup4==4.4.0
Django==1.8.5
django-appconf==1.0.1
django-compressor==1.5
django-extensions==1.5.6
django-mathfilters==0.3.0
django-modelcluster==0.6.2
django-redis==4.2.0
django-taggit==0.17.1
django-treebeard==3.0
djangorestframework==3.2.3
html5lib==0.999
msgpack-python==0.4.6
mysqlclient==1.3.6
Pillow==2.9.0
pytz==2015.4
redis==2.10.3
six==1.9.0
Unidecode==0.4.18
wagtail==1.1
werkzeug
https://github.com/bloodywing/Willow/archive/master.zip
